+++ sketch/program.js
@@ -1,8 +1,8 @@
 class Program {
-  constructor(lightArray, interval) {
+  constructor(lightMatrix, interval) {
     this.defaultLightColor = color(200, 140, 5);
 
-    this.lights = lightArray;
+    this.lights = lightMatrix;
 
     this.isActive = false;
 
@@ -13,59 +13,54 @@ class Program {
   }
 
   update() {
-    this.lights.forEach(function(light) {
-      light.update();
-    });
+
   }
 
-  _lightUpCurrent() {
-    this.lights[this.currentLightIndex].turnOn(this.defaultLightColor);
+  _lightUpRow(index, color) {
+    this.lights.lightUpRow(index, color)
   }
 
-  _turnOffCurrrent() {
-    this.lights[this.currentLightIndex].turnOff();
+  _turnOffRow(index) {
+    this.lights.turnOffRow(index);
   }
 
-  _fadeCurrent() {
-    this.lights[this.currentLightIndex].fade(this.fadeTime);
+  _fadeRow(index, fadeTime) {
+    this.lights.fadeRow(index, fadeTime)
   }
 
-  _next() {
-    this.currentLightIndex++;
-    if (this.currentLightIndex >= this.lights.length) {
-      this.currentLightIndex = 0;
+  _lightUpCol(index, color) {
+    this.lights.lightUpCol(index, color)
+  }
+
+  _turnOffCol(index) {
+    this.lights.turnOffCol(index);
+  }
+
+  _fadeCol(index, fadeTime) {
+    this.lights.fadeCol(index, fadeTime)  }
+
+    _isTimeForUpdate() {
+      return (millis() - this.lastChange) > this.interval;
     }
-  }
 
-  _previous() {
-    this.currentLightIndex--;
-    if (this.currentLightIndex < 0) {
-      this.currentLightIndex = this.lights.length -1;
+    start() {
+      this.isActive = true;
     }
-  }
 
-  _isTimeForUpdate() {
-    return (millis() - this.lastChange) > this.interval;
-  }
+    stop() {
+      this.isActive = false;
+    }
 
-  start() {
-    this.isActive = true;
-  }
+    // _dimAll() {
+    //   this.lights.forEach(function(light) {
+    //     light.dim(this.dimAmount);
+    //   }, this);
+    // }
+    //
+    // _turnOffAll() {
+    //   this.lights.forEach(function(light) {
+    //     light.turnOff();
+    //   });
+    // }
 
-  stop() {
-    this.isActive = false;
   }
-
-  _dimAll() {
-    this.lights.forEach(function(light) {
-      light.dim(this.dimAmount);
-    }, this);
-  }
-
-  _turnOffAll() {
-    this.lights.forEach(function(light) {
-      light.turnOff();
-    });
-  }
-
-}