+++ CHANGELOG.md
@@ -5,9 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
+
+## [0.5.0] - 2018-08-26
 ### Added
 - Lights handle their own fading.
 - Light coordination separated into programs, e.g. Pulse, Reverse Pulse.
+- Lights organised as matrix instead of array.
+- Matrix pulse both horizontal and vertical.
+
+### Changed
+- Project name: Light Array -> Lights. To accommodate matrix configuration.
 
 ## [0.4.0] - 2018-08-25
 ### Added